commands: use a partial to simplify the mode argument
This uses a partial from the functools module to simplify decorating
commands, even further than the MODE global did, by not requiring the
argument at all. This reduces a bit of boilerplate, combined with a
shorter function name this allows for some registrations to fit on a
single line that didn't before.